chore(settlement): rename variables to address review comments
