+++ crates/interledger-settlement/src/lib.rs
@@ -1,4 +1,4 @@
 // export the API only when explicitly asked
 #[cfg(feature = "settlement_api")]
-pub mod settlement_api;
-pub mod settlement_core;
+pub mod api;
+pub mod core;
